+++ packages/engine/src/avatar/AvatarAnimationSystem.ts
@@ -29,7 +29,6 @@ import {
 } from '../transform/components/DistanceComponents'
 import { TransformComponent } from '../transform/components/TransformComponent'
 import { updateGroupChildren } from '../transform/systems/TransformSystem'
-import { XRHand } from '../xr/XRComponents'
 import { getControlMode, XRState } from '../xr/XRState'
 import { updateAnimationGraph } from './animation/AnimationGraph'
 import { solveLookIK } from './animation/LookAtIKSolver'
@@ -47,6 +46,8 @@ import {
 import { AvatarHeadDecapComponent } from './components/AvatarIKComponents'
 import { AvatarHeadIKComponent } from './components/AvatarIKComponents'
 import { LoopAnimationComponent } from './components/LoopAnimationComponent'
+import { applyInputSourcePoseToIKTargets } from './functions/applyInputSourcePoseToIKTargets'
+import { updateLocalAvatarPosition, updateLocalAvatarRotation } from './functions/moveAvatar'
 
 export const AvatarAnimationState = defineState({
   name: 'AvatarAnimationState',
@@ -57,9 +58,6 @@ export const AvatarAnimationState = defineState({
 
 const _vector3 = new Vector3()
 const _vec = new Vector3()
-const _rotXneg30 = new Quaternion().setFromAxisAngle(new Vector3(1, 0, 0), -Math.PI * 0.3)
-const _rotY90 = new Quaternion().setFromAxisAngle(new Vector3(0, 0, 1), Math.PI / 2)
-const _rotYneg90 = new Quaternion().setFromAxisAngle(new Vector3(0, 0, 1), -Math.PI / 2)
 
 export default async function AvatarAnimationSystem(world: World) {
   await AnimationManager.instance.loadDefaultAnimations()
@@ -102,7 +100,6 @@ export default async function AvatarAnimationSystem(world: World) {
   const filterPriorityEntities = (entity: Entity) =>
     world.priorityAvatarEntities.has(entity) || entity === world.localClientEntity
 
-  const xrState = getState(XRState)
   const filterFrustumCulledEntities = (entity: Entity) =>
     !(
       DistanceFromCameraComponent.squaredDistance[entity] > minimumFrustumCullDistanceSqr &&
@@ -114,83 +111,24 @@ export default async function AvatarAnimationSystem(world: World) {
   let sortedTransformEntities = [] as Entity[]
 
   const execute = () => {
-    const { localClientEntity, elapsedSeconds, deltaSeconds } = world
-    const xrFrame = Engine.instance.xrFrame!
+    const { elapsedSeconds, deltaSeconds } = world
 
-    const inAttachedControlMode = getControlMode() === 'attached'
+    /**
+     * 1 - Update local client movement
+     */
 
-    const referenceSpace = getState(XRState).originReferenceSpace.value!
+    updateLocalAvatarPosition()
+    updateLocalAvatarRotation()
 
-    /** Update controller pose input sources from WebXR into the ECS */
-    if (xrFrame && referenceSpace && hasComponent(localClientEntity, AvatarIKTargetsComponent)) {
-      /** Head */
-      if (inAttachedControlMode && hasComponent(localClientEntity, AvatarHeadIKComponent)) {
-        const ik = getComponent(localClientEntity, AvatarHeadIKComponent)
-        ik.target.quaternion.copy(world.camera.quaternion)
-        ik.target.position.copy(world.camera.position)
-      }
+    /**
+     * 2 - Calculate localClientEntity's IK targets
+     */
 
-      for (const inputSource of world.inputSources) {
-        /** Left Hand */
-        if (inputSource.handedness === 'left' && hasComponent(localClientEntity, AvatarLeftHandIKComponent)) {
-          const ik = getComponent(localClientEntity, AvatarLeftHandIKComponent)
-          const hand = inputSource.hand as XRHand | undefined
-          /** detect hand joint pose support */
-          if (hand && xrFrame.getJointPose) {
-            const wrist = hand.get('wrist')
-            if (wrist) {
-              const jointPose = xrFrame.getJointPose(wrist, referenceSpace)
-              if (jointPose) {
-                ik.target.position.copy(jointPose.transform.position as unknown as Vector3)
-                ik.target.quaternion.copy(jointPose.transform.orientation as unknown as Quaternion)
-                ik.target.quaternion.multiply(_rotYneg90) // @todo look into this
-              }
-            }
-          } else if (inputSource.gripSpace) {
-            const pose = Engine.instance.xrFrame!.getPose(inputSource.gripSpace, referenceSpace)
-            if (pose) {
-              ik.target.position.copy(pose.transform.position as any as Vector3)
-              ik.target.quaternion.copy(pose.transform.orientation as any as Quaternion)
-            }
-          } else {
-            const pose = Engine.instance.xrFrame!.getPose(inputSource.targetRaySpace, referenceSpace)
-            if (pose) {
-              ik.target.position.copy(pose.transform.position as any as Vector3)
-              ik.target.quaternion.copy(pose.transform.orientation as any as Quaternion)
-            }
-          }
-        }
+    applyInputSourcePoseToIKTargets()
 
-        /** Right Hand */
-        if (inputSource.handedness === 'right' && hasComponent(localClientEntity, AvatarRightHandIKComponent)) {
-          const ik = getComponent(localClientEntity, AvatarRightHandIKComponent)
-          const hand = inputSource.hand as XRHand | undefined
-          if (hand && xrFrame.getJointPose) {
-            const wrist = hand.get('wrist')
-            if (wrist) {
-              const jointPose = xrFrame.getJointPose(wrist, referenceSpace)
-              if (jointPose) {
-                ik.target.position.copy(jointPose.transform.position as unknown as Vector3)
-                ik.target.quaternion.copy(jointPose.transform.orientation as unknown as Quaternion)
-                ik.target.quaternion.multiply(_rotY90) // @todo look into this
-              }
-            }
-          } else if (inputSource.gripSpace) {
-            const pose = Engine.instance.xrFrame!.getPose(inputSource.gripSpace, referenceSpace)
-            if (pose) {
-              ik.target.position.copy(pose.transform.position as any as Vector3)
-              ik.target.quaternion.copy(pose.transform.orientation as any as Quaternion)
-            }
-          } else {
-            const pose = Engine.instance.xrFrame!.getPose(inputSource.targetRaySpace, referenceSpace)
-            if (pose) {
-              ik.target.position.copy(pose.transform.position as any as Vector3)
-              ik.target.quaternion.copy(pose.transform.orientation as any as Quaternion)
-            }
-          }
-        }
-      }
-    }
+    /**
+     * 3 - Sort & apply avatar priority queue
+     */
 
     let needsSorting = false
     avatarSortAccumulator += deltaSeconds
@@ -224,6 +162,10 @@ export default async function AvatarAnimationSystem(world: World) {
     }
 
     priorityQueue.update()
+
+    /**
+     * 4 - Apply avatar animations
+     */
 
     const avatarAnimationEntities = avatarAnimationQuery(world).filter(filterPriorityEntities)
     const headIKEntities = headIKQuery(world).filter(filterPriorityEntities)
@@ -290,6 +232,10 @@ export default async function AvatarAnimationSystem(world: World) {
       const rootPos = AnimationManager.instance._defaultRootBone.position
       rig.Hips.position.setX(rootPos.x).setZ(rootPos.z)
     }
+
+    /**
+     * 5 - Apply avatar IK
+     */
 
     /**
      * Apply head IK
@@ -400,11 +346,6 @@ export default async function AvatarAnimationSystem(world: World) {
      * we need to manually do it for Loop Animation Entities
      */
     for (const entity of loopAnimationEntities) updateGroupChildren(entity)
-
-    /**
-     * Update threejs skeleton manually
-     *  - overrides default behaviour in WebGLRenderer.render, calculating mat4 multiplcation
-     */
   }
 
   const cleanup = async () => {
